Make SequenceJsonProvider usable with both ILoggingEvent and IAccessEvent
Refactor SequenceJsonProvider and associate utility classes to make them usable for both ILoggingEvent and IAccessEvent:
- create a new SequenceJsonProvider by making a copy of the existing SequenceJsonProvider (restricted to ILoggingEvent)
- deprecate the existing SequenceJsonProvider in favour of the new common SequenceJsonProvider
- move LoggingEventJsonProviders#addSequence() method into the common JsonProviders class

Closes #661
